requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
matplotlib==3.9.2
Pillow==11.0.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.62.3
python-dotenv==0.19.1
Flask==3.1.0
flask-cors==5.0.0
